Add common package for use between packages
