+++ packages/ai_insight_worker/package.json
@@ -18,6 +18,7 @@
   },
   "dependencies": {
     "@prisma/client": "4.8.1",
+    "common": "link:../common",
     "dotenv-cli": "^6.0.0",
     "openai": "^4.78.1",
     "prisma": "4.8.1",
