--- packages/ai_insight_worker/src/types/index.ts
@@ -1,9 +1,0 @@
-type Task = {
-  projectId: number;
-  sprintId: number;
-  user: string;
-};
-
-export type {
-  Task
-};
