+++ packages/backend/package.json
@@ -37,6 +37,7 @@
     "@types/multer": "^1.4.7",
     "axios": "^1.3.4",
     "bcryptjs": "^2.4.3",
+    "common": "link:../common",
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv-cli": "^6.0.0",
