+++ packages/common/README.md
@@ -1,0 +1,1 @@
+TROFOS common package, if different services have coupling, like redis keys/ channels
