+++ packages/common/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "common",
+  "version": "1.0.0",
+  "description": "",
+  "main": "src/index.ts",
+  "packageManager": "pnpm@10.4.1"
+}
